+++ netpyne/specs/simConfig.py
@@ -47,6 +47,7 @@ class SimConfig (object):
         self.gatherOnlySimData = False  # omits gathering of net+cell data thus reducing gatherData time
         self.compactConnFormat = False  # replace dict format with compact list format for conns (need to provide list of keys to include)
         self.connRandomSecFromList = True  # select random section (and location) from list even when synsPerConn=1 
+        self.distributeSynsUniformly = True  # locate synapses at uniformly across section list; if false, place one syn per section in section list   
         self.saveCellSecs = True  # save all the sections info for each cell (False reduces time+space; available in netParams; prevents re-simulation)
         self.saveCellConns = True  # save all the conns info for each cell (False reduces time+space; prevents re-simulation)
         self.timing = True  # show timing of each process